requirements.txt
asgiref==3.2.10
certifi==2020.4.5.2
chardet==3.0.4
Django==2.1.5
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-g-recaptcha==0.2.4
django-ip==1.0.2
django-session-timeout==0.1.0
django-six==1.0.4
django-tz-detect==0.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
funcsigs==1.0.2
idna==2.9
mock==4.0.2
pbr==5.5.0
Pillow==8.1.0
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
sweetify==2.0.0
urllib3==1.25.9